+++ core-ng/src/main/java/core/framework/api/module/CDNConfig.java
@@ -1,0 +1,24 @@
+package core.framework.api.module;
+
+import core.framework.impl.module.ModuleContext;
+
+/**
+ * @author neo
+ */
+public final class CDNConfig {
+    private final ModuleContext context;
+
+    public CDNConfig(ModuleContext context) {
+        this.context = context;
+    }
+
+    public void hosts(String... hosts) {
+        context.httpServer.siteManager.templateManager.cdnHosts(hosts);
+    }
+
+    public void version(String version) {
+        String cdnVersion = version;
+        if (version.startsWith("${")) cdnVersion = "local";
+        context.httpServer.siteManager.templateManager.cdnVersion(cdnVersion);
+    }
+}